// bidly/apps/server/src/app/open-api/bid/types/bid-etc.type.ts
export type Bid_Etc = {
  bidNtceNo: string;
  bidNtceOrd: string;
  untyNtceNo: string;
  rgstTyNm: string;
  ntceKindNm: string;
  bidNtceDt: string;
  refNo: string;
  bidNtceNm: string;
  ntceInsttCd: string;
  ntceInsttNm: string;
  dminsttNm: string;
  bidMethdNm: string;
  cntrctCnclsMthdNm: string;
  sucsfbidMthdNm: string;
  bidBeginDt: string;
  bidClseDt: string;
  opengDt: string;
  ntceSpecDocUrl1: string;
  ntceSpecDocUrl2: string;
  ntceSpecDocUrl3: string;
  ntceSpecDocUrl4: string;
  ntceSpecDocUrl5: string;
  ntceSpecFileNm1: string;
  ntceSpecFileNm2: string;
  ntceSpecFileNm3: string;
  ntceSpecFileNm4: string;
  ntceSpecFileNm5: string;
  rbidPermsnYn: string;
  presmptPrce: string;
  opengPlce: string;
  bidNtceDtlUrl: string;
  bidNtceUrl: string;
  bidPrtcptFee: string;
  bidPrtcptFeePaymntYn: string;
  cmmnSpldmdYn: string;
  rgstDt: string;
  bidGrntymnyPaymntObjYn: string;
  bidQlfctRgstCntnts: string;
  rmrkCntnts: string;
};